% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANsearcher.R
\name{CRANsearcher}
\alias{CRANsearcher}
\title{CRANsearcher}
\usage{
CRANsearcher()
}
\description{
Addin for searching packages in CRAN database based on keywords
}
\examples{
\dontrun{
CRANsearcher()
}

}
