% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANsearcher.R
\name{CRANsearcher}
\alias{CRANsearcher}
\title{CRANsearcher}
\usage{
CRANsearcher(viewer = NULL)
}
\arguments{
\item{viewer}{Shiny app viewer options. Default is \code{dialogViewer("Search packages...", width = 1200, height = 900)}.}
}
\description{
Addin for searching packages in CRAN database based on keywords
}
\examples{
\dontrun{
CRANsearcher()
}

}
